requirements.txt
google-cloud-logging>=3.11.2
google-cloud-monitoring>=2.20.0
google-cloud-dataproc>=5.7.0
google-cloud-aiplatform>=1.63.0
pyyaml>=6.0.1
pydantic>=2.8.2
